requirements.txt
opencv-python
mediapipe
python-vlc